backend/requirements.txt
# ============================================================================
# FINANCE APP RAG - Python Dependencies
# ============================================================================
# Актуально на: январь 2026
# Python: 3.11+
# ============================================================================

# ===== Core Framework =====
fastapi==0.115.6             
uvicorn[standard]==0.34.0     
python-multipart==0.0.18      

# ===== Database & ORM =====
sqlalchemy==2.0.36            
alembic==1.14.0               

# ===== Security & Auth =====
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4          
python-dotenv==1.0.1              

# ===== Data Validation =====
pydantic==2.10.5           

# ===== Vector Database & Embeddings =====
qdrant-client==1.12.1         
fastembed==0.4.2            
           
